Removed the CSS fade-in transition effect for the overlay, since it made the goggles feel sluggish, especially on more complex pages. This also simplified the code a bit.
Added an animation effect to mixmaster remix, including undo/redo, to make it easier to see where changes are taking place. This is especially useful for people who want to scrub through a hacked page's history to observe what's been changed.

The implementation of the latter is not ideal, since it's putting presentation logic in the ReplaceWithCmd logic, which will cause "tear off this page" to generate weird visual effects. Will fix this soon.