The touch toolbar is now always visible, even on non-touch devices (it should probably be renamed to just 'toolbar' later). On non-touch devices, though, only the quit, undo/redo, and publish features are shown.
This change comes from the Hive Tokyo sprint on Jan 22, where the kids were learning both Scratch and Hackasaurus, and ended up embedding their Scratch animation/game into a webpage with the goggles. Unfortunately, this sometimes resulted in the Scratch Java Applet swallowing all the keypresses, which rendered the kids unable to publish or further remix their hacks.

Aside from helping with cases in which keyboard focus is lost, of course, this also helps with recognition vs. recall.